requirements.txt
streamlit
feedparser
pandas
google-generativeai
duckduckgo-search